:bug: post 요청을 통해서 중복 회원가입 검사를 하도록 수정
